+++ lib/acts_as_tenant/model_extensions.rb
@@ -46,7 +46,7 @@ module ActsAsTenant
         before_validation proc { |m|
           if ActsAsTenant.current_tenant
             if options[:polymorphic]
-              m.send(:"#{fkey}=", ActsAsTenant.current_tenant.class.to_s) if m.send(fkey.to_s).nil?
+              m.send(:"#{fkey}=", ActsAsTenant.current_tenant.send(pkey)) if m.send(fkey.to_s).nil?
               m.send(:"#{polymorphic_type}=", ActsAsTenant.current_tenant.class.to_s) if m.send(polymorphic_type.to_s).nil?
             else
               m.send :"#{fkey}=", ActsAsTenant.current_tenant.send(pkey)